requirements.txt
azure-functions
azure-storage-blob>=12.14.0
pandas>=1.5.3
psycopg2-binary>=2.9.5
azure-identity>=1.12.0
azure-keyvault-secrets>=4.6.0
python-dotenv>=1.0.0
retrying>=1.3.4
certifi>=2023.7.22